requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.2.2
Flask-Mail==0.9.1
gunicorn==20.0.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.6
Werkzeug==2.2.2
zipp==3.11.0
pymongo==4.3.3
pandas==1.5.1
altair==4.2.2
dash==2.9.3
dash_bootstrap_components==1.4.1
plotly==5.11.0
vega_datasets==0.9.0
bokeh==3.0.3
python-dotenv==1.0.0
pytest==7.3.1
mongomock
kafka-python==2.0.2